Move functions to the right place for the deprecation level
