+++ gtk/application.go
@@ -118,70 +118,6 @@ func (v *Application) SetMenubar(m *glib.MenuModel) {
 	C.gtk_application_set_menubar(v.native(), mptr)
 }
 
-// GetMenuByID is a wrapper around gtk_application_get_menu_by_id().
-func (v *Application) GetMenuByID(id string) *glib.Menu {
-	cstr1 := (*C.gchar)(C.CString(id))
-	defer C.free(unsafe.Pointer(cstr1))
-
-	c := C.gtk_application_get_menu_by_id(v.native(), cstr1)
-	if c == nil {
-		return nil
-	}
-	return &glib.Menu{glib.MenuModel{wrapObject(unsafe.Pointer(c))}}
-}
-
-// ListActionDescriptions is a wrapper around gtk_application_list_action_descriptions().
-func (v *Application) ListActionDescriptions() []string {
-	var descs []string
-	c := C.gtk_application_list_action_descriptions(v.native())
-	originalc := c
-	defer C.g_strfreev(originalc)
-
-	for *c != nil {
-		descs = append(descs, C.GoString((*C.char)(*c)))
-		c = C.next_gcharptr(c)
-	}
-
-	return descs
-}
-
-// GetAccelsForAction is a wrapper around gtk_application_get_accels_for_action().
-func (v *Application) GetAccelsForAction(act string) []string {
-	cstr1 := (*C.gchar)(C.CString(act))
-	defer C.free(unsafe.Pointer(cstr1))
-
-	var descs []string
-	c := C.gtk_application_get_accels_for_action(v.native(), cstr1)
-	originalc := c
-	defer C.g_strfreev(originalc)
-
-	for *c != nil {
-		descs = append(descs, C.GoString((*C.char)(*c)))
-		c = C.next_gcharptr(c)
-	}
-
-	return descs
-}
-
-// SetAccelsForAction is a wrapper around gtk_application_set_accels_for_action().
-func (v *Application) SetAccelsForAction(act string, accels []string) {
-	cstr1 := (*C.gchar)(C.CString(act))
-	defer C.free(unsafe.Pointer(cstr1))
-
-	caccels := C.make_strings(C.int(len(accels) + 1))
-	defer C.destroy_strings(caccels)
-
-	for i, accel := range accels {
-		cstr := C.CString(accel)
-		defer C.free(unsafe.Pointer(cstr))
-		C.set_string(caccels, C.int(i), (*C.gchar)(cstr))
-	}
-
-	C.set_string(caccels, C.int(len(accels)), nil)
-
-	C.gtk_application_set_accels_for_action(v.native(), cstr1, caccels)
-}
-
 // IsInhibited is a wrapper around gtk_application_is_inhibited().
 func (v *Application) IsInhibited(flags ApplicationInhibitFlags) bool {
 	return gobool(C.gtk_application_is_inhibited(v.native(), C.GtkApplicationInhibitFlags(flags)))
